# speed up `int_pctl.tune_results()`