@simonpcouch 2024-03-04
Now:

``` r
set.seed(1)
bench::mark(
  int_res_new2 = int_pctl(c5_res)
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 × 6
#>   expression        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 int_res_new2    2.99s    2.99s     0.335     430MB     10.0
```

<sup>Created on 2024-03-04 with [reprex v2.1.0](https://reprex.tidyverse.org)</sup>